+++ runtime/font-manage.h
@@ -51,11 +51,11 @@ CharModelSet *FontManager_loadFont(FontManager *manager, const Font *font);
 CharModelSet *FontManager_findFont(FontManager *manager, const Font *font);
 Array/*<Vertex2D>*/ *
 charModelSetGenHCoordArray(const CharModelSet *set, const Array /*<char_t>*/ *char_array,
-                           const Vertex2D * anchor, int32_t c_space, uint32_t mode,
+                           const Vertex2D * anchor, float c_space, uint32_t mode,
                            XGLVector2D feedback_vec, const Allocator *allocator);
 Array/*<Vertex2D>*/ *
 charModelSetGenVCoordArray(const CharModelSet *set, const Array /*<char_t>*/ *char_array,
-                           const Vertex2D * anchor, int32_t c_space, uint32_t mode,
+                           const Vertex2D * anchor, float c_space, uint32_t mode,
                            XGLVector2D feedback_vec, const Allocator *allocator);
 
 #endif  // XIDE_FONT_MANAGE_H
